third_party/WebKit/LayoutTests/media/activation-behavior-accesskey.html

<!doctype html>
<title>activation behavior using accesskey</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<video controls accesskey="x"></video>
<script>
function pressAccessKey(key)
{
    if (navigator.userAgent.search(/\bMac OS X\b/) != -1)
        modifiers = ["ctrlKey", "altKey"];
    else
        modifiers = ["altKey"];
    eventSender.keyDown(key, modifiers);
}

test(function()
{
    var v = document.querySelector("video");
    assert_true(v.paused, "paused state before access key press");
    pressAccessKey("x");
    assert_false(v.paused, "paused state before access key press");
});
</script>
